refactor(frontend): consolidate api utility proxies and error handlers
