--- frontend/src/app/api/_utils/analytics.ts
@@ -1,12 +1,0 @@
-/**
- * Server-side proxy handler for the analytics endpoint
- */
-export async function proxyAnalytics() {
-	const API_URL = process.env.AZURE_FUNCTION_URL;
-
-	const response = await fetch(`${API_URL}/analytics`);
-	if (!response.ok) {
-		throw new Error(`Failed to fetch analytics: ${response.statusText}`);
-	}
-	return response.json();
-}
