--- frontend/src/app/api/_utils/errorHandler.test.ts
@@ -1,41 +1,0 @@
-import { beforeEach, describe, expect, it, vi } from "vitest";
-import { handleApiError } from "./errorHandler";
-
-describe("errorHandler", () => {
-	beforeEach(() => {
-		vi.spyOn(console, "error").mockImplementation(() => {});
-	});
-
-	describe("handleApiError", () => {
-		it("returns 500 status with error message for Error instance", () => {
-			const error = new Error("Test error message");
-			const response = handleApiError(error, "testing");
-
-			expect(response.status).toBe(500);
-		});
-
-		it("includes error message in unified error format", async () => {
-			const errorMessage = "Custom error";
-			const error = new Error(errorMessage);
-			const response = handleApiError(error, "testing");
-
-			const body = await response.json();
-			expect(body.error).toBe(errorMessage);
-			expect(body).not.toHaveProperty("success");
-		});
-
-		it("handles non-Error objects with generic message", async () => {
-			const response = handleApiError("String error", "testing");
-
-			const body = await response.json();
-			expect(body.error).toBe("An unexpected error occurred");
-		});
-
-		it("sets correct content-type header", () => {
-			const error = new Error("Test error");
-			const response = handleApiError(error, "testing");
-
-			expect(response.headers.get("content-type")).toBe("application/json");
-		});
-	});
-});
