--- frontend/src/app/api/_utils/health.ts
@@ -1,16 +1,0 @@
-export interface HealthResponse {
-	status: string;
-	timestamp: string;
-}
-
-/**
- * Server-side proxy handler for the health endpoint
- */
-export async function proxyHealth() {
-	const API_URL = process.env.AZURE_FUNCTION_URL;
-	const response = await fetch(`${API_URL}/health`);
-	if (!response.ok) {
-		throw new Error(`Health check failed: ${response.statusText}`);
-	}
-	return response.json();
-}
